+++ netket/operator/__init__.py
@@ -1,16 +1,17 @@
-from .local_values import local_values
+from ._local_values import local_values
 
-from .der_local_values import der_local_values
+from ._der_local_values import der_local_values
 
 
-from .local_operator import LocalOperator
-from .local_liouvillian import LocalLiouvillian
-from .graph_operator import GraphOperator
+from ._local_operator import LocalOperator
+from ._local_liouvillian import LocalLiouvillian
+from ._graph_operator import GraphOperator
 
 from . import spin, boson
 
-from .hamiltonian import Ising, Heisenberg
+from ._hamiltonian import Ising, Heisenberg
 
-from .abstract_operator import AbstractOperator
-from .bose_hubbard import BoseHubbard
-from .pauli_strings import PauliStrings
+from ._abstract_operator import AbstractOperator
+from ._bose_hubbard import BoseHubbard
+from ._pauli_strings import PauliStrings
+
