+++ netket/operator/_abstract_operator.py
