+++ netket/operator/_local_values.py
@@ -1,7 +1,7 @@
 import numpy as _np
 from numba import jit
 
-from .local_liouvillian import LocalLiouvillian as _LocalLiouvillian
+from ._local_liouvillian import LocalLiouvillian as _LocalLiouvillian
 from netket.machine.density_matrix.abstract_density_matrix import (
     AbstractDensityMatrix as DensityMatrix,
 )
