+++ netket/operator/spin.py
@@ -1,4 +1,4 @@
-from .local_operator import LocalOperator
+from ._local_operator import LocalOperator
 
 import numpy as _np
